terraform: add stateadd to its own test file
